+++ src/angular/video.component.ts
@@ -6,7 +6,11 @@ import {
     Output,
     ViewEncapsulation,
 } from "@angular/core";
-import type { Anchor, Mode, Placeholder, StateEvent } from "../_/types";
+// eslint-disable-next-line no-duplicate-imports
+import type { OnChanges } from "@angular/core";
+import { parseDuration, parseFrom, parseTo } from "../_/parse";
+import { preComputeVideoOptions } from "../_/preCompute";
+import type { Anchor, Mode, Placeholder, StateEvent, VideoOptions } from "../_/types";
 
 @Component( {
     "selector": `TwicVideo`,
@@ -30,6 +34,7 @@ import type { Anchor, Mode, Placeholder, StateEvent } from "../_/types";
             [transitionDelay]="transitionDelay"
             [transitionDuration]="transitionDuration"
             [transitionTimingFunction]="transitionTimingFunction"
+            [videoOptions]="videoOption"
             (stateChangeEvent)="onStateChange($event)"
         ></TwicMedia>
     `,
@@ -39,26 +44,42 @@ import type { Anchor, Mode, Placeholder, StateEvent } from "../_/types";
         "class": `twic-i twic-d`,
     },
 } )
-export class TwicVideoComponent {
+export class TwicVideoComponent implements OnChanges {
     mediaTag = `video`;
     @Input() anchor: Anchor = undefined;
     @Input() bot: string = undefined;
+    @Input() duration: number | string = undefined;
     @Input() focus: string = undefined;
+    @Input() from: number | string = undefined;
     @Input() intrinsic: string = undefined;
     @Input() mode: Mode = undefined;
     @Input() eager: boolean | string;
     @Input() placeholder: Placeholder = undefined;
     @Input() position: string = undefined;
+    @Input() posterFrom: number | string = undefined;
     @Input() preTransform: string = undefined;
     @Input() ratio: number | string = undefined;
     @Input() src: string;
-    @Input() step: number = undefined;
+    @Input() step: number | string = undefined;
     @Input() title: string = undefined;
+    @Input() to: number | string = undefined;
     @Input() transition:boolean | string;
     @Input() transitionDelay: string = undefined;
     @Input() transitionDuration: string = undefined;
     @Input() transitionTimingFunction: string = undefined;
     @Output() stateChangeEvent = new EventEmitter< StateEvent >();
+    _duration: number;
+    _from: number;
+    _posterFrom: number;
+    _to: number;
+    videoOption: VideoOptions;
+    ngOnChanges( ): void {
+        this._duration = parseDuration( this.duration );
+        this._from = parseFrom( this.from );
+        this._posterFrom = parseDuration( this.posterFrom );
+        this._to = parseTo( this.to );
+        this.videoOption = preComputeVideoOptions( this._duration, this._from, this._posterFrom, this._to );
+    }
     onStateChange( stateEvent: StateEvent ) {
         this.stateChangeEvent.emit( stateEvent );
     }